requirements.txt
scipy==1.0.1
numpy==1.14.2
matplotlib==2.1.1
pysoundfile==0.9.0